Bit.{test_any,test_all,popcount}: Fix wrong datatype

+++ src/lua/bit.cpp
@@ -134,15 +134,15 @@ namespace
 	});
 
 	function_ptr_luafun lua_testany(LS, "bit.test_any", [](lua_state& L, const std::string& fname) -> int {
-		uint64_t a = L.get_numeric_argument<uint8_t>(1, fname.c_str());
-		uint64_t b = L.get_numeric_argument<uint8_t>(2, fname.c_str());
+		uint64_t a = L.get_numeric_argument<uint64_t>(1, fname.c_str());
+		uint64_t b = L.get_numeric_argument<uint64_t>(2, fname.c_str());
 		L.pushboolean((a & b) != 0);
 		return 1;
 	});
 
 	function_ptr_luafun lua_testall(LS, "bit.test_all", [](lua_state& L, const std::string& fname) -> int {
-		uint64_t a = L.get_numeric_argument<uint8_t>(1, fname.c_str());
-		uint64_t b = L.get_numeric_argument<uint8_t>(2, fname.c_str());
+		uint64_t a = L.get_numeric_argument<uint64_t>(1, fname.c_str());
+		uint64_t b = L.get_numeric_argument<uint64_t>(2, fname.c_str());
 		L.pushboolean((a & b) == b);
 		return 1;
 	});
@@ -160,7 +160,7 @@ namespace
 	}
 
 	function_ptr_luafun lua_popcount(LS, "bit.popcount", [](lua_state& L, const std::string& fname) -> int {
-		uint64_t a = L.get_numeric_argument<uint8_t>(1, fname.c_str());
+		uint64_t a = L.get_numeric_argument<uint64_t>(1, fname.c_str());
 		L.pushnumber(popcount(a));
 		return 1;
 	});
